add patient tests

+++ theralogs/views/patient.py
@@ -1,9 +1,9 @@
 from django.contrib.auth.decorators import login_required
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponse
 from django.shortcuts import redirect, render
 
-from theralogs.models import Patient
 from theralogs.forms import EditPatientForm, NewPatientForm
+from theralogs.models import Patient
 
 
 @login_required
@@ -30,8 +30,13 @@ def create_patient(request):
                 )
                 patient.save()
                 return redirect("profile")
-        context = {"form": form}
-        return render(request, "theralogs/patient/new_patient.html", context)
+            else:
+                response = render(
+                    request, "theralogs/patient/new_patient.html", {"form": form}
+                )
+                response.status_code = 400
+                return response
+        return render(request, "theralogs/patient/new_patient.html", {"form": form})
     return redirect("profile")
 
 
@@ -46,15 +51,22 @@ def edit_patient(request, patient_id):
             patient.email = form.cleaned_data.get("email")
             patient.save()
             return redirect("profile")
-    context = {"form": form}
-    return render(request, "theralogs/patient/edit_patient.html", context)
+        else:
+            response = render(
+                request, "theralogs/patient/edit_patient.html", {"form": form}
+            )
+            response.status_code = 400
+            return response
+    return render(request, "theralogs/patient/edit_patient.html", {"form": form})
 
 
 @login_required
 def delete_patient(request, patient_id):
-    patient = Patient.objects.get(id=patient_id)
-    if request.method == "POST":
-        patient.delete()
-        return redirect("profile")
-    context = {"patient_name": patient.name}
-    return render(request, "theralogs/patient/delete_patient.html", context)
+    patient = Patient.objects.filter(id=patient_id).first()
+    if patient:
+        if request.method == "POST":
+            patient.delete()
+            return redirect("profile")
+        context = {"patient_name": patient.name}
+        return render(request, "theralogs/patient/delete_patient.html", context)
+    return HttpResponse("Patient not found", status=400)
